Fix: PARENT_OF failed to get parent Argument
Argument.interface now not a pointer and owns the TypeInterface
instance.

This solution is just one of the ways to make PARENT_OF work correctly.
PARENT_OF macro just requires address of the 'instance' field, which we
could have passed, but it being a pointer, meant involving pointer to
pointer things and on the other hand may be its just a good idea for
Argument to own the things it requires.