Removed TypeInterface.to_string interface function
This function which converted the argument 'type' back to string is not
required when the default value is given as string. This change opened
led to other changes for simplification like removing the 'is_optional'
argument from 'argument_add' function etc. This change also thus made
implementing a custom TypeInterface easier (there is one less function
to implement).