cargs_add_arg now only takes string literal
* 'name', 'description' and 'default_value' are now assumed to be string
  literals and they are validated for that. Since they all now have
  static lifetimes they are no longer copied.

The rational for this change is that all setting up of arguments and
actual parsing of command line args, is done together and then is never
used again, therefore string literal does not need to be copied as the
scope does not change between setup and parsing.

Note: String literals must be passed directly and not a pointer to them,
this is because is difficult to verify pointers.

Also in this commit:
* CARGS_ARGUMENT_PREFIX_CHAR_OVERRIDE now need to be a string quoted char.